discord_bot/requirements.txt
python-dotenv
boto3
requests
pynacl
